ec2bot-open/app/requirements.txt
python-telegram-bot>=20.0
boto3>=1.26.0
pyyaml>=6.0
